refactor(sdk): Move timeline m.room_key handler into new module
