+++ crates/matrix-sdk/src/room/timeline/mod.rs
@@ -29,7 +29,7 @@ use ruma::{
     events::{fully_read::FullyReadEventContent, relation::Annotation, AnyMessageLikeEventContent},
     EventId, OwnedEventId, OwnedUserId, TransactionId, UInt,
 };
-use tracing::{debug_span, error, instrument};
+use tracing::{error, instrument};
 
 use super::{Joined, Room};
 use crate::{
@@ -43,9 +43,10 @@ mod event_item;
 mod inner;
 #[cfg(test)]
 mod tests;
+#[cfg(feature = "e2e-encryption")]
+mod to_device;
 mod virtual_item;
 
-use self::inner::TimelineInner;
 pub use self::{
     event_item::{
         EncryptedMessage, EventTimelineItem, Message, ReactionDetails, Sticker, TimelineDetails,
@@ -53,6 +54,7 @@ pub use self::{
     },
     virtual_item::VirtualTimelineItem,
 };
+use self::{inner::TimelineInner, to_device::handle_room_key_event};
 
 /// A high-level view into a regular¹ room's contents.
 ///
@@ -110,54 +112,9 @@ impl Timeline {
         // Not using room.add_event_handler here because RoomKey events are
         // to-device events that are not received in the context of a room.
         #[cfg(feature = "e2e-encryption")]
-        let room_id = room.room_id().to_owned();
-        #[cfg(feature = "e2e-encryption")]
-        let room_key_handle = room.client.add_event_handler({
-            use std::iter;
-
-            use ruma::events::room_key::ToDeviceRoomKeyEvent;
-            use tracing::{trace, Instrument};
-
-            use crate::Client;
-
-            let inner = inner.clone();
-            move |event: ToDeviceRoomKeyEvent, client: Client| {
-                let inner = inner.clone();
-                let room_id = room_id.clone();
-                async move {
-                    if event.content.room_id != room_id {
-                        let event_room_id = &event.content.room_id;
-                        let session_id = &event.content.session_id;
-                        trace!(
-                            %event_room_id, timeline_room_id = %room_id, %session_id,
-                            "Received to-device room key event for a different room, ignoring"
-                        );
-                        return;
-                    }
-
-                    let Some(olm_machine) = client.olm_machine() else {
-                        error!("The olm machine isn't yet available");
-                        return;
-                    };
-
-                    let session_id = event.content.session_id;
-                    let Some(own_user_id) = client.user_id() else {
-                        error!("The user's own ID isn't available");
-                        return;
-                    };
-
-                    inner
-                        .retry_event_decryption(
-                            &room_id,
-                            olm_machine,
-                            iter::once(session_id.as_str()).collect(),
-                            own_user_id,
-                        )
-                        .await;
-                }
-                .instrument(debug_span!("handle_to_device_room_key_event"))
-            }
-        });
+        let room_key_handle = room
+            .client
+            .add_event_handler(handle_room_key_event(inner.clone(), room.room_id().to_owned()));
         #[cfg(feature = "e2e-encryption")]
         let _room_key_handler_guard = room.client.event_handler_drop_guard(room_key_handle);
